apps/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
fastapi
uvicorn
python-multipart
PyJWT
passlib[bcrypt]
slowapi
cryptography
pyotp
qrcode[pil]
